Fix lake not generated properly (#29)
Potentially fix something else :) Instances order matters now, wasn't
the case previously. oGame should be the last instance added to
a room, otherwise room generation will not work properly.
The same issue occurred with the Olmec room (fixed, but it was
a part of the a0789ac commit, so it was hard to track the solution).